Move virtuaBranch.ts into branches folder
