+++ apps/desktop/src/components/PullRequestPreview.svelte
@@ -4,10 +4,10 @@
 	import Markdown from '$components/Markdown.svelte';
 	import { BaseBranchService } from '$lib/baseBranch/baseBranchService';
 	import { BranchController } from '$lib/branches/branchController';
+	import { VirtualBranchService } from '$lib/branches/virtualBranchService';
 	import { showError } from '$lib/notifications/toasts';
 	import { Project } from '$lib/project/project';
 	import { RemotesService } from '$lib/remotes/remotesService';
-	import { VirtualBranchService } from '$lib/vbranches/virtualBranch';
 	import { getContext } from '@gitbutler/shared/context';
 	import Badge from '@gitbutler/ui/Badge.svelte';
 	import Button from '@gitbutler/ui/Button.svelte';
