+++ apps/desktop/src/routes/[projectId]/+layout.ts
@@ -4,6 +4,7 @@ import { BranchController } from '$lib/branches/branchController';
 import { BranchListingService } from '$lib/branches/branchListing';
 import { BranchDragActionsFactory } from '$lib/branches/dragActions.js';
 import { GitBranchService } from '$lib/branches/gitBranch';
+import { VirtualBranchService } from '$lib/branches/virtualBranchService';
 import { CommitDragActionsFactory } from '$lib/commits/dragActions.js';
 import { CommitService } from '$lib/commits/service';
 import { StackingReorderDropzoneManagerFactory } from '$lib/dragging/stackingReorderDropzoneManager';
@@ -18,7 +19,6 @@ import { TemplateService } from '$lib/pr/templateService';
 import { type Project } from '$lib/project/project';
 import { ProjectService } from '$lib/project/projectService';
 import { UpstreamIntegrationService } from '$lib/vbranches/upstreamIntegrationService';
-import { VirtualBranchService } from '$lib/vbranches/virtualBranch';
 import { error } from '@sveltejs/kit';
 import type { LayoutLoad } from './$types';
